added seperate styles file to job screen
